requirements.txt
matplotlib==3.5.2
numpy==1.21.5
pandas==1.4.4
Pillow==9.3.0
requests==2.28.1
scikit_learn==1.2.0
seaborn==0.11.2
streamlit==1.16.0
streamlit_lottie==0.0.3
